requirements.txt
flask
python-novaclient
